+++ CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated Stencil to v1.2.5
 - Changed the event name for the `manifold-auth-token` component from the stencil auto-generated name to `manifold-token-receive` and documented that event.
 
+### Fixed
+
+- Fixed issue where success event of `manifold-data-rename-resource` was emitted before renamed resource was ready.
+
 ## [v0.5.3]
 
 ### Fixed